+++ rollup.all.config.mjs
@@ -1,0 +1,4 @@
+import highlighterConfig from './rollup.config.mjs';
+import contentConfig from './rollup.content.config.mjs';
+
+export default [highlighterConfig, contentConfig];